Extract AppRunningSpeed into it's own component
