+++ src/pages/ExecDashboard/index.js
@@ -1,4 +1,5 @@
-import { SampleBreakdownBlock } from './SampleBreakdownBlock';
+import AppRunningSpeed from './AppRunningSpeed';
+import SampleBreakdownBlock from './SampleBreakdownBlock';
 import React from 'react';
 import classnames from 'classnames';
 import Loader from '../../components/Loader';
@@ -7,12 +8,6 @@ import './ExecutiveDashboard.scss';
 import apiData from '../../apiData';
 import Spinner from '../../components/Spinner';
 import SamplesProcessedBlock from './SamplesProcessedBlock';
-
-const RunningStatus = {
-  NotRunning: 0,
-  Running: 1,
-  RunningFast: 2
-};
 
 class ExecutiveDashboard extends React.Component {
   render() {
@@ -31,7 +26,7 @@ class ExecutiveDashboard extends React.Component {
             // Executive dashboard
             return (
               <>
-                <AppRunningSpeed speed={getRunningSpeed(data)} />
+                <AppRunningSpeed data={data} />
 
                 <div className="exec-dash__grid">
                   <div className="exec-dash__block">
@@ -69,27 +64,6 @@ class ExecutiveDashboard extends React.Component {
 }
 
 export default ExecutiveDashboard;
-
-function AppRunningSpeed({ speed }) {
-  const title =
-    speed === RunningStatus.RunningFast
-      ? 'We’re processing data at super sonic speed!'
-      : speed === RunningStatus.Running
-        ? 'We’re processing data'
-        : 'We’re not processing data';
-
-  return (
-    <div
-      className={classnames('running-status', {
-        'running-status--not-running': speed === RunningStatus.NotRunning,
-        'running-status--running': speed === RunningStatus.Running,
-        'running-status--running-fast': speed === RunningStatus.RunningFast
-      })}
-    >
-      {title}
-    </div>
-  );
-}
 
 async function fetchStats() {
   return {
@@ -140,22 +114,4 @@ async function fetchStats() {
   };
 
   return await Ajax.get('/stats');
-}
-
-/**
- * Returns how fast the system is running
- * ref https://github.com/AlexsLemonade/refinebio-frontend/issues/453#issuecomment-445031974
- * @param {*} stats stats object as returned by `/stats`
- */
-function getRunningSpeed(stats) {
-  const activeVolumes = stats.active_volumes.length;
-  const runningNomadJobs = parseInt(stats.nomad_running_jobs, 10);
-
-  if (activeVolumes > 5 && runningNomadJobs > 200) {
-    return RunningStatus.RunningFast;
-  } else if (activeVolumes > 5 && runningNomadJobs > 50) {
-    return RunningStatus.Running;
-  } else {
-    return RunningStatus.NotRunning;
-  }
 }
